fixes after pull request comment in seprating validation logic
