+++ pom.xml
@@ -122,6 +122,18 @@
         <artifactId>commons-validator</artifactId>
         <version>1.6</version>
     </dependency>
+    
+    <dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger2</artifactId>
+			<version>2.4.0</version>
+		</dependency>
+
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger-ui</artifactId>
+			<version>2.4.0</version>
+		</dependency>
 
 
 	</dependencies>
