+++ src/main/java/org/bahmni/insurance/service/AOpernmrsFhirConstructorService.java
@@ -1,10 +1,9 @@
 package org.bahmni.insurance.service;
 
+import java.io.IOException;
+
 import org.bahmni.insurance.ImisConstants;
 import org.bahmni.insurance.client.RestTemplateFactory;
-import org.hl7.fhir.dstu3.model.Claim;
-import org.hl7.fhir.dstu3.model.EligibilityRequest;
-import org.hl7.fhir.dstu3.model.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
@@ -19,13 +18,13 @@ public abstract class AOpernmrsFhirConstructorService {
 		return restFactory.getRestTemplate(ImisConstants.OPENMRS_FHIR);
 	}
 
-	public abstract Claim constructFhirClaimRequest(String insuranceID);
+	public abstract String constructFhirClaimRequest(String insuranceID)throws IOException;
 
-	public abstract EligibilityRequest constructFhirEligibilityRequest(String insuranceID);
+	public abstract String constructFhirEligibilityRequest(String insuranceID) throws IOException;
 	
 
 	public abstract String getFhirPatient(String patientId);
 
-	public abstract Task constructFhirClaimTrackRequest(String insuranceID);
+	public abstract String constructFhirClaimTrackRequest(String insuranceID) throws IOException;
 
 }
